backend/requirements.txt
asgiref==3.5.0
astroid==2.11.2
asttokens==2.0.5
backcall==0.2.0
country-converter==0.7.4
decorator==5.1.1
dill==0.3.4
Django==4.0.3
django-cors-headers==3.12.0
djangorestframework==3.13.1
executing==0.8.3
ipython==8.3.0
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.7.0
numpy==1.22.4
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.5.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
Pygments==2.12.0
pylint==2.13.4
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.2.0
typing_extensions==4.1.1
wcwidth==0.2.5
wrapt==1.14.0
whitenoise